+++ SeedShop/src/main/java/com/apu/seedshop/repository/AnOrderRepository.java
@@ -9,8 +9,8 @@ import org.springframework.data.repository.CrudRepository;
 import com.apu.seedshop.jpa.AnOrder;
 import org.springframework.data.repository.query.Param;
 
-public interface AnOrderRepository extends CrudRepository<AnOrder, Integer> {
-    public List<AnOrder> findById(@Param("id") Integer id);
+public interface AnOrderRepository extends CrudRepository<AnOrder, Long> {
+    public List<AnOrder> findById(@Param("id") Long id);
     @Override
     public List<AnOrder> findAll();
 }
